refactor: move `access` services file
- Move `access` services to `services/access/access.ts`.
- Reflect structure of `access.test.ts` test file.
- Update file paths in `access` controller and its tests.
